Refactored into Player class
Now just create a Player and give it the html page, input directory, and
output directory, and a Player will fetch the page, find the game logs
for all of them, then save the scores, stddev, etc in its fields.

Also saves to a csv at the end of this then loads from it subsequently.

envsetup.sh sets up the PYTHONPATH